+++ libs/aws/sabi-s3/src/core/client.rs
@@ -1,6 +1,5 @@
-use crate::core::S3Bucket;
+use crate::core::{S3Bucket, S3Result};
 use crate::operations;
-use crate::S3Result;
 
 #[derive(Debug)]
 pub struct S3Client {
