+++ libs/aws/sabi-s3/src/core/mod.rs
@@ -3,3 +3,5 @@ pub use bucket::S3Bucket;
 
 mod client;
 pub use client::S3Client;
+
+pub use super::error::Result as S3Result;
