+++ libs/aws/sabi-s3/src/operations/put_object/error.rs
@@ -1,7 +1,5 @@
 extern crate failure;
 
-pub type Result<T> = ::std::result::Result<T, Error>;
-
 #[derive(Fail, Debug)]
 pub enum Error {
     #[fail(display = "FileNotFound > {}", 0)]
